Update rename data.py to models.py
