+++ resources/meals.py
@@ -5,7 +5,7 @@ import datetime
 from flask import jsonify, Blueprint, make_response
 from flask_restful import Resource, Api, reqparse, inputs
 
-import data
+import models as data
 
 
 class MealList(Resource):
